Create SSO
If you are trying to integrate with Ecwid and you are running in .NET framework, you can use this code and deploy them on your website which contain the Ecwid widgets like Product widgets and it should create the Ecwid customer and log the user in. On the second visit the customer is just signed in to you Ecwid.